Add forc, forc-plugins, fuel-core and fuel-gql-cli pkgs with devShell
This adds the following packages to the flake:

- forc
- forc-explore
- forc-fmt
- forc-lsp
- fuel-core
- fuel-gql-cli

Each package is currently built from a pinned instance of the master
branch of their respective repo and is not associated with a particular
version. In the future, the aim is to allow for selecting any git tag
version.

Currently the latest stable Rust release provided by oxalica's
rust-overlay is used (1.63 as of writing this). This can be updated if
necessary with `nix flake lock --update-input rust-overlay`.

A `devShell` has also been added that combines all of the inputs for the
packages above under a single shell for ease of spinning up a
development environment.