Update rust-overlay for latest Rust stable 1.64 release
It looks like my previous rust-overlay update in #23 was a little early
and didn't include 1.64 yet. #23 didn't show any issues yet as none of
the packages at the time of that commit were released after 2022-09-23,
and so we didn't hit the error until the latest nightlies merged into
`master` with #27 which also didn't show any issues as it was was opened
before #23 was merged :')

This should fix it! It's probably worth enabling the same branch
protection we have on the Sway repo where merging a PR requires it's up
to date with `master` to avoid cases like this.