Add required nix configuration to flake `nixConfig` attribute
An attempt at enabling the fuellabs nix cache via the flake's
`nixConfig` attribute. The idea is to hopefully remove the need for
adding cache info to the user's global config and ease usage slightly.

From `nix help flake`:

> nixConfig: a set of nix.conf options to be set when evaluating any
> part of a flake. In the interests of security, only a small set of
> whitelisted options (currently bash-prompt, bash-prompt-prefix,
> bash-prompt-suffix, and flake-registry) are allowed to be set without
> confirmation so long as accept-flake-config is not set in the global
> configuration.

Currently this doesn't seem to be working - I'm running into the same
issue described here https://github.com/NixOS/nix/issues/6672. Either
the user has to be part of the "trusted-users" set (i.e. is allowed to
specify new substituters) or the fuellabs cache already has to be part
of the global `trusted-substituters` list, otherwise the fuellabs cache
is ignored.