Fix bug in `fuel-dev` due to `forc-explore-nightly` not yet being available. Check `nix develop` in CI. (#39)
No commits have been made to forc-explore since including adding it to
the Nix flake - as a result, no nightly package has been made available.
This makes the forc-explore-nightly inclusion optional and dependent on
whether or not a nightly actually exists.

This slipped through as during CI we'd only build the packages, however
the `devShell`s were never checked. This adds another job to the PR CI
that checks the `fuel-dev` `devShell` on both macOS and Linux. Also adds
a similar step to the refresh-manifests CI to catch any unexpected
breakage there too.

Also closes #38.