Add a `fuel.nix` mdbook (#61)
This adds a new `book/` directory containing an mdbook for `fuel.nix`.

The new book will replace the README with a much friendlier "Quick
Start", and a more searchable and detailed guide covering how to install
Nix, use fuel packages, and how to contribute to fuel.nix itself.

Closes #59.

TODO
----

- [x] Add a new "Quick Start" section with the new two-step installation
approach.
- [x] Merge existing README content into subchapters.
- [x] Add a more detailed "Nix Setup" chapter for those interested in
more details.
- [x] Add a section on `nix profile install` for persistent installation
to the "Packages" chapter.
- [x] Add a new "Contributing" section with details on how to add/update
packages and milestones.
- [x] Update the README to simply point to the book and Quick Start.
- [x] Add a CI step to publish the book.
- ~Open an issue to remove the **NOTE:** under the `nix profile install`
section required due to DeterminateSystems/nix-installer#477.~ It seems
this issue only occurs if a previous Nix installation existed but was
not cleaned up properly. The `note` has been updated.