ci: Refresh and commit manifest files for patch PRs (#135)
Rather than having to run refresh-manifests locally when making a patch
PR, this change will run the refresh script in CI and commit the
generated manifests to the patch PR.

It only runs on PRs that include changes to `patches.nix`. This will
ensure that we don't forget to include these files. The script runs
father in CI than locally, so it's easier all around.

Example commit on this PR:
https://github.com/FuelLabs/fuel.nix/pull/135/commits/545201664986ddf5cb05b7ef124af498710c50b8

---------

Co-authored-by: GitHub Action <action@github.com>